Before implementing new genome
Added unitary tests using pytest, the folder structures has been changed for adding the tests.

Altruism now works with relatedness based on a benefit-relatedness ratio. It also has maximum cost sacrificed, which will be the cost in the majority of cases, minimum cost for unrelated inds and benefit to cost ratio

The population is now structured in groups that have migration between them in each generation, all functions have been changed to work with this structure.